+++ components/layouts/Header/HeaderBar.tsx
@@ -1,15 +1,18 @@
+import { Breadcrumbs } from '../../core/Breadcrumbs/Breadcrumbs';
 import { CustomLink } from '../../core/CustomLink/CustomLink';
 import { StyledHeaderBarWrapper, StyledHeaderBar } from './styles';
+import { BreadcrumbsData } from '../../core/types';
 
 type HeaderBarProps = {
-  onClickBack: () => void;
+  onClickBreadcrumb: (route: string, asRoute: string) => void;
   onClickFavorites: () => void;
+  breadcrumbs: BreadcrumbsData[];
 };
 
-export const HeaderBar: React.FC<HeaderBarProps> = ({ onClickBack, onClickFavorites }) => (
+export const HeaderBar: React.FC<HeaderBarProps> = ({ onClickBreadcrumb, onClickFavorites, breadcrumbs }) => (
   <StyledHeaderBarWrapper data-testid="header-bar">
     <StyledHeaderBar className="grid-container">
-      <CustomLink onClick={onClickBack}>{'< BACK'}</CustomLink>
+    <Breadcrumbs breadcrumbsData={breadcrumbs || []} onClickBreadcrumb={onClickBreadcrumb} />
       <CustomLink onClick={onClickFavorites}>
         <i className="material-icons">bookmarks</i>FAVORITES
       </CustomLink>
